backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
numpy==1.24.3
onnxruntime==1.17.1

